move test utils to utils file
