+++ src/http/api/v1/matches.ts
@@ -105,7 +105,7 @@ export const createHandler = (config: HttpApiConfig): express.Router => {
     };
 
     const calculator = new Calculator(calculatorOptions);
-    const matches = await calculator.calculate();
+    const matches = await calculator.calculate(defaultProportionalMatchOptions);
     const responseBody = JSON.stringify(matches, (_key, value) =>
       typeof value === "bigint" ? value.toString() : (value as unknown)
     );
