+++ src/test/http/app.test.ts
@@ -2,63 +2,25 @@
 import { describe, test, expect, beforeEach, vi } from "vitest";
 import express from "express";
 import os from "os";
-import fs from "fs/promises";
-import path from "path";
 import request, { Response as SupertestResponse } from "supertest";
 import { createHttpApi } from "../../http/app.js";
-import {
-  AugmentedResult,
-  DataProvider,
-  FileNotFoundError,
-} from "../../calculator/index.js";
+import { AugmentedResult, DataProvider } from "../../calculator/index.js";
 import { PriceProvider } from "../../prices/provider.js";
 import { Logger } from "pino";
 import { PotentialVotes } from "../../http/api/v1/matches.js";
-import { PassportScore } from "../../passport/index.js";
 import { Chain } from "../../config.js";
 import { constants } from "ethers";
-import { Price } from "../../prices/common.js";
+import {
+  TestPriceProvider,
+  TestPassportProvider,
+  TestDataProvider,
+  loadFixture,
+} from "../utils.js";
 
 vi.spyOn(os, "hostname").mockReturnValue("dummy-hostname");
 
 // Typed version of supertest's Response
 type Response<T> = Omit<SupertestResponse, "body"> & { body: T };
-
-const loadFixture = async (
-  name: string,
-  extension = "json"
-): Promise<string> => {
-  const p = path.resolve(__dirname, "../fixtures", `${name}.${extension}`);
-  const data = await fs.readFile(p, "utf8");
-  return data;
-};
-
-type Fixtures = { [path: string]: string | undefined | unknown[] };
-
-export class TestPriceProvider {
-  async convertToUSD() {
-    return Promise.resolve({ amount: 0 });
-  }
-
-  async convertFromUSD() {
-    return Promise.resolve({ amount: 0 });
-  }
-
-  async getUSDConversionRate(
-    chainId: number,
-    tokenAddress: string,
-    _blockNumber?: number
-  ): Promise<Price & { decimals: number }> {
-    return Promise.resolve({
-      price: 1_000_000_000,
-      token: tokenAddress,
-      block: 0,
-      code: "",
-      decimals: 18,
-      timestamp: 0,
-    });
-  }
-}
 
 const MOCK_CHAINS = [
   {
@@ -71,44 +33,6 @@ const MOCK_CHAINS = [
     ],
   },
 ] as Chain[];
-
-class TestPassportProvider {
-  _fixture: PassportScore[] | null = null;
-
-  async start() {}
-
-  async stop() {}
-
-  async getScoreByAddress(address: string): Promise<PassportScore | undefined> {
-    if (this._fixture === null) {
-      this._fixture = JSON.parse(
-        await loadFixture("passport_scores")
-      ) as PassportScore[];
-    }
-    return this._fixture.find((score) => score.address === address);
-  }
-}
-
-export class TestDataProvider implements DataProvider {
-  fixtures: Fixtures;
-
-  constructor(fixtures: Fixtures) {
-    this.fixtures = fixtures;
-  }
-
-  async loadFile<T>(description: string, path: string): Promise<Array<T>> {
-    const fixture = this.fixtures[path];
-    if (fixture === undefined) {
-      throw new FileNotFoundError(description);
-    }
-
-    if (typeof fixture !== "string") {
-      return fixture as Array<T>;
-    }
-
-    return JSON.parse(await loadFixture(fixture)) as Array<T>;
-  }
-}
 
 const DUMMY_LOGGER = {
   debug: () => {},
